+++ src/components/ReviewBox.js
@@ -1,10 +1,10 @@
-export default function CaixaComentario() {
+export default function ReviewBox() {
     // Esse componente deve conter todos os elemntos da caixa de comentarios.
     // Alem de mudar dinamicamnte caso o usuario logado no momento ja tenha 
     // feito um comentario, adidionando assim os elementos de exclusao e edicao
     return (
         <p>
-            Hello caixa de CaixaComentario
+            Hello caixa de ReviewBox
         </p>
     );
 }
