+++ src/components/SearchBox.js
@@ -1,4 +1,4 @@
-export default function CaixaBusca() {
+export default function SearchBox() {
     // Esse componente deve ter a caixa de busca que deve mudar de forma. 
     // Deve ser um icone de busca quando o usuario nao estiver pesquisando 
     // nada mas ao ser clicado deve se tranformar em um input. Alem disso
