+++ src/pages/ConfigScreen.js
@@ -1,4 +1,4 @@
-export default function TelaConfiguracoes() {
+export default function ConfigScreen() {
     // Essa tela tem duas versoes a no modo de edicao e 
     // a no modo de visualizacao, ver prototipo no figma
     return (
