+++ src/pages/MovieDetailScreen.js
@@ -1,0 +1,19 @@
+import SearchBox from "../components/SearchBox";
+import ReviewBox from "../components/ReviewBox";
+import MoviesBox from "../components/MoviesBox"; // Para os Semelhantes
+
+export default function MovieDetailScreen() {
+    return (
+        <>
+            <h1>
+                Hello tela de detalhe do filmes
+            </h1>
+            <SearchBox />
+            <MoviesBox />
+            <ReviewBox />
+            <ReviewBox />
+            <ReviewBox />
+            <ReviewBox />
+        </>
+    );
+}
