+++ src/pages/RegistrationScreen.js
@@ -1,4 +1,4 @@
-export default function TelaCadasto() {
+export default function RegistrationScreen() {
     return (
         <h1>
             Hello tela de cadastro
