+++ src/pages/SeachScreen.js
@@ -1,0 +1,14 @@
+import SearchBox from "../components/SearchBox";
+import MoviesBox from "../components/MoviesBox";
+
+export default function SeachScreen() {
+    return (
+        <>
+            <h1>
+                Hello tela de busca
+            </h1>
+            <SearchBox />
+            <MoviesBox />
+        </>
+    );
+}
