--- src/pages/TelaBusca.js
@@ -1,14 +1,0 @@
-import CaixaBusca from "../components/CaixaBusca";
-import CaixaFilmes from "../components/CaixaFilmes";
-
-export default function TelaBusca() {
-    return (
-        <>
-            <h1>
-                Hello tela de busca
-            </h1>
-            <CaixaBusca />
-            <CaixaFilmes />
-        </>
-    );
-}
