--- src/pages/TelaUsuario.js
@@ -1,17 +1,0 @@
-import CaixaFilmes from "../components/CaixaFilmes";
-
-export default function TelaUsuario() {
-    return (
-        <>
-            <h1>
-                Hello tela de usuario
-            </h1>
-
-            <CaixaFilmes />
-            <CaixaFilmes />
-            <CaixaFilmes />
-
-        </>
-
-    );
-}
