+++ src/pages/UserScreen.js
@@ -1,0 +1,17 @@
+import MoviesBox from "../components/MoviesBox";
+
+export default function UserScreen() {
+    return (
+        <>
+            <h1>
+                Hello tela de usuario
+            </h1>
+
+            <MoviesBox />
+            <MoviesBox />
+            <MoviesBox />
+
+        </>
+
+    );
+}
